+++ docs/changelog.rst
@@ -27,6 +27,13 @@ New features
   * New decorator :py:meth:`botogram.Bot.chat_unavailable`
   * New method :py:meth:`botogram.Component.add_chat_unavailable_hook`
 
+* Added support for editing messages sent by the bot
+
+  * New method :py:meth:`botogram.Message.edit`
+  * New method :py:meth:`botogram.Message.edit_caption`
+  * New method :py:meth:`botogram.Bot.edit_message`
+  * New method :py:meth:`botogram.Bot.edit_caption`
+
 * Every method which sends something to a chat now returns the sent
   :py:class:`~botogram.Message`